Updated the Template Engine benchmarks
