+++ Intercode.Toolbox.TemplateEngine.Benchmarks/Program.cs
@@ -2,15 +2,17 @@
 // Author:      Eduardo Velasquez
 // Copyright (c) 2024, Intercode Consulting, Inc.
 
+using System.Globalization;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Exporters.Csv;
 using BenchmarkDotNet.Reports;
 using BenchmarkDotNet.Running;
 using Intercode.Toolbox.TemplateEngine.Benchmarks;
+using Perfolizer.Horology;
 using Perfolizer.Metrology;
 
 #if PROFILING
-var tests = new MacroProcessingTests();
+var tests = new MacroProcessingBenchmarks();
 
 for( int i = 0; i < 100; ++i )
 {
@@ -23,16 +25,16 @@ return;
 var exporter = new CsvExporter(
   CsvSeparator.CurrentCulture,
   new SummaryStyle(
-    cultureInfo: System.Globalization.CultureInfo.CurrentCulture,
+    cultureInfo: CultureInfo.CurrentCulture,
     printUnitsInHeader: true,
     printUnitsInContent: false,
-    timeUnit: Perfolizer.Horology.TimeUnit.Nanosecond,
+    timeUnit: TimeUnit.Nanosecond,
     sizeUnit: SizeUnit.B
   ));
 
 var config = ManualConfig.CreateMinimumViable().AddExporter(exporter);
 
-BenchmarkSwitcher.FromAssembly( typeof( MacroProcessingTests ).Assembly )
+BenchmarkSwitcher.FromAssembly( typeof( MacroProcessingBenchmarks ).Assembly )
                  .Run( args );
 
 #endif
